+++ src/NSDate+Convenience.h
@@ -10,17 +10,17 @@
 
 @interface NSDate (Convenience)
 
-- (int)year;
-- (int)month;
-- (int)day;
-- (int)hour;
+- (NSInteger)year;
+- (NSInteger)month;
+- (NSInteger)day;
+- (NSInteger)hour;
 - (NSString *)weekString;
-- (NSDate *)offsetDay:(int)numDays;
+- (NSDate *)offsetDay:(NSInteger)numDays;
 - (BOOL)isToday;
 
-+ (NSDate *)dateForDay:(unsigned int)day month:(unsigned int)month year:(unsigned int)year;
++ (NSDate *)dateForDay:(NSInteger)day month:(NSInteger)month year:(NSInteger)year;
 + (NSDate *)dateStartOfDay:(NSDate *)date;
-+ (int)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate;
++ (NSInteger)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate;
 + (NSDate *)dateFromString:(NSString *)dateString format:(NSString *)format;
 + (NSString *)stringFromDate:(NSDate *)date format:(NSString *)format;
 + (NSDate *)dateFromString:(NSString *)dateString;
