+++ src/NSDate+Convenience.m
@@ -8,7 +8,7 @@
 
 @implementation NSDate (Convenience)
 
-- (int)year {
+- (NSInteger)year {
     NSCalendar *gregorian = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSYearCalendarUnit fromDate:self];
@@ -16,28 +16,28 @@
 }
 
 
-- (int)month {
+- (NSInteger)month {
     NSCalendar *gregorian = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSMonthCalendarUnit fromDate:self];
     return [components month];
 }
 
-- (int)day {
+- (NSInteger)day {
     NSCalendar *gregorian = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSDayCalendarUnit fromDate:self];
     return [components day];
 }
 
-- (int)hour {
+- (NSInteger)hour {
     NSCalendar *gregorian = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [gregorian components:NSHourCalendarUnit fromDate:self];
     return [components hour];
 }
 
-- (NSDate *)offsetDay:(int)numDays {
+- (NSDate *)offsetDay:(NSInteger)numDays {
     NSCalendar *gregorian = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
 //    [gregorian setFirstWeekday:2]; //monday is first day
@@ -56,7 +56,7 @@
     return [[NSDate dateStartOfDay:self] isEqualToDate:[NSDate dateStartOfDay:[NSDate date]]];
 }
 
-+ (NSDate *)dateForDay:(unsigned int)day month:(unsigned int)month year:(unsigned int)year
++ (NSDate *)dateForDay:(NSInteger)day month:(NSInteger)month year:(NSInteger)year
 {
     NSCalendar *gregorian = [[NSCalendar alloc] initWithCalendarIdentifier:NSGregorianCalendar];
     NSDateComponents *components = [[NSDateComponents alloc] init];
@@ -122,13 +122,13 @@
     return @"";
 }
 
-+ (int)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate {
++ (NSInteger)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate {
     NSCalendar *calendar = [[NSCalendar alloc]
             initWithCalendarIdentifier:NSGregorianCalendar];
     unsigned int unitFlags = NSDayCalendarUnit;
     NSDateComponents *comps = [calendar components:unitFlags fromDate:startDate toDate:endDate options:0];
     //    int months = [comps month];
-    int days = [comps day];
+    NSInteger days = [comps day];
     return days;
 }
